comporasion_project/requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
PyPDF2>=3.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0 